Week 15

+++ bmimetric2.cpp
@@ -1,6 +1,6 @@
 #include "iostream"
 
-int bmimetricf(int weight, float height) {
+float bmimetricf(int weight, float height) {
     float data = weight / (height * height);
     if (data < 18.5) {
         std::cout << "Underweight";
